source/services/datavault-backup/requirements.txt
redshift_connector==2.0.918
boto3-stubs[secretsmanager] # boto3 local typing
requests==2.31.0
